create package menu and moved there menu classes
