+++ src/main/java/com/infoshareacademy/App.java
@@ -1,5 +1,6 @@
 package com.infoshareacademy;
 
+import com.infoshareacademy.menu.MenuLogin;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -33,7 +34,9 @@ public class App {
             STDOUT.info(temp);
         }
         STDOUT.info(holidaysJsonData.toString());
-        MenuHoliday menuHoliday = new MenuHoliday();
-        menuHoliday.menuHoliday();
+        
+        MenuLogin menuLogin = new MenuLogin();
+        menuLogin.Login();
+
     }
 }
